main,refactor: make parsers.h main part private
