+++ main/parse.c
@@ -28,7 +28,7 @@
 #include "main.h"
 #include "param.h"
 #include "parse_p.h"
-#include "parsers.h"
+#include "parsers_p.h"
 #include "promise.h"
 #include "promise_p.h"
 #include "ptag_p.h"
