+++ win32/ctags_vs2013.vcxproj
@@ -281,7 +281,7 @@
     <ClInclude Include="..\main\param.h" />
     <ClInclude Include="..\main\parse.h" />
     <ClInclude Include="..\main\parse_p.h" />
-    <ClInclude Include="..\main\parsers.h" />
+    <ClInclude Include="..\main\parsers_p.h" />
     <ClInclude Include="..\main\portable-dirent_p.h" />
     <ClInclude Include="..\main\promise.h" />
     <ClInclude Include="..\main\promise_p.h" />
